start unsplitting
