--- src/dra_psp/EC60.c
@@ -1,69 +1,0 @@
-// SPDX-License-Identifier: AGPL-3.0-or-later
-#include "../dra/dra.h"
-
-extern s32 g_StatBuffTimers[16];
-
-s32 CalcAttack(s32 equipId, u32 otherEquipId) {
-    s32 i;
-    s16 equipmentAttackBonus;
-    s16 totalAttack;
-    s16 strengthStat;
-
-    if (g_EquipDefs[equipId].itemCategory == ITEM_FOOD ||
-        g_EquipDefs[equipId].itemCategory == ITEM_MEDICINE) {
-        return 0;
-    }
-    if (g_EquipDefs[equipId].itemCategory == ITEM_SHIELD &&
-        g_EquipDefs[equipId].attack == 1) {
-        return 0;
-    }
-
-    if (equipId == ITEM_ALUCARD_SHIELD) {
-        return 0;
-    }
-
-    equipmentAttackBonus = 0;
-
-    for (i = 0; i < 5; i++) {
-        equipmentAttackBonus +=
-            g_AccessoryDefs[g_Status.equipment[2 + i]].attBonus;
-    }
-
-    totalAttack = g_EquipDefs[equipId].attack;
-    strengthStat = g_Status.statsTotal[0];
-
-    if (totalAttack <= strengthStat) {
-        totalAttack += strengthStat;
-    } else {
-        totalAttack += strengthStat / 2;
-    }
-
-    totalAttack += equipmentAttackBonus;
-
-    if (equipId == ITEM_BADELAIRE) {
-        totalAttack += g_Status.timerHours;
-    }
-    if (equipId == ITEM_MURAMASA) {
-        totalAttack += SquareRoot0(g_Status.D_80097C40);
-    }
-    if (equipId == 4 && g_EquipDefs[otherEquipId].itemCategory == ITEM_SHIELD) {
-        totalAttack += 5;
-    }
-    if (equipId == ITEM_SWORD_FAMILIAR) {
-        totalAttack += g_Status.statsFamiliars[FAM_STATS_SWORD].level;
-    }
-    if (g_StatBuffTimers[1]) {
-        totalAttack += 20;
-    }
-    if (totalAttack < 0) {
-        totalAttack = 0;
-    }
-    if (totalAttack > 999) {
-        totalAttack = 999;
-    }
-    return totalAttack;
-}
-
-INCLUDE_ASM("dra_psp/psp/dra_psp/EC60", func_psp_090EB850);
-
-INCLUDE_ASM("dra_psp/psp/dra_psp/EC60", func_psp_090EB8B8);
